+++ app/controllers/CourseController.php
@@ -1,33 +1,29 @@
 <?php
-//    This file is part of Alternate Admin for Moodle
-//    Alternate Admin Free GNU Application
-//    Copyright (C) 2021 Manuel Gil.
+// This file is part of Moodle - http://moodle.org/
 //
-//    This program is free software: you can redistribute it and/or modify
-//    it under the terms of the GNU General Public License as published by
-//    the Free Software Foundation, either version 3 of the License, or
-//    (at your option) any later version.
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
 //
-//    This program is distributed in the hope that it will be useful,
-//    but WITHOUT ANY WARRANTY; without even the implied warranty of
-//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-//    GNU General Public License for more details.
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
 //
-//    You should have received a copy of the GNU General Public License
-//    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
 /**
  * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  * $Id$
  * <p>Title: Alternate Admin for Moodle.</p>
  * <p>Description: This wrapper for Moodle adds a new interface to
  * 					streamline your administrative tasks.</p>
- * <p>Copyright: <a href="http://www.gnu.org/copyleft/gpl.html">GNU GPL v3 or later</a>.</p>
- * <p>Company: <a href="https://imgil.dev/">Manuel Gil</a></p>
  *
- * Problem: Add more function to tradiccional admin.
- * @author $Author: Manuel Gil. $
- * @version $Revision: 0.3.1 $ $Date: 02/16/2021 $
+ * @package		wrapper
+ * @author 		$Author: 2021 Manuel Gil. <https://imgil.dev/> $
+ * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  */
 
@@ -50,7 +46,7 @@ class CourseController extends BaseController
 	 */
 	public function anyIndex()
 	{
-		header('location: ' . BASE_URL);
+		redirect(BASE_URL);
 	}
 
 	/**